Implemented LevelModeProgressLine.cs; Fixed GetProgress method in LevelTaskCompletionChecker.cs; Fixed Construct method in BaseProgressLine.cs.

+++ Assets/Code/Scripts/Gameplay/LevelTaskCompletionChecker.cs
@@ -45,7 +45,7 @@ namespace Core.Gameplay
 
         public float GetProgress()
         {
-            return _completedTaskElementCount / _totalTaskElementCount;
+            return (float)_completedTaskElementCount / _totalTaskElementCount;
         }
 
         private void HandleCellExplosion(CellExplosionResult cellExplosionResult)
